Add bitarray callback
If there is a function that can add more bytes into the buffer, it is
possible to register it. The function will be called whenever bitarray
calls reach the end of the buffer. It is possible to rely on this for
the first buffer allocation too. There is a default handler so it can
be called unconditionally. For example:

        static int input_func(struct bitarray *b)
        {
                static unsigned char data[] = { 1, 2 };

                b->data = data;
                b->size = 2;
                return 0;
        }

        ...
                bitarray_init(&b, NULL, 0);
                bitarray_callback(&b, input_func);

                for (i = 0; i < 8; i++)
                        printf("%ld", bitarray_fetch(&b, 8));
                fputs("\n", stdout);
        ...

This must print "12121212\n" to standard output if CHAR_BIT == 8.