Add huffman_table function
This function does not have a self-explanatory interface. The caller
gives an unused array of unsigned ints and the function puts pairs of
bit-lengths and reversed Huffman code values into it. Although this
description and the code itself may look complex, the usage of should
be quite straightforward. Later commits will introduce code that uses
this function and all the pieces of the puzzle should fit together.