Introduce dy-zip compression
The dy-zip program can compress files but it is not very efficient. I
think the compression ratio could be acceptable but the program itself
is too slow. It is not a surprise because the current implementation
is very simple and relies on "brute force". However, it should be easy
(or at least more pleasant) to improve the algorithms because I have a
working version available.