Change chunk size
If there were many small files, the program would allocate a lot of
memory. I know that this strategy is not very good but it is not that
simple to get file sizes in a portable way. I will think about adding
some non-portable code (preprocessor ifdefs) because that would make
this part of file I/O more efficient. However, usually this is not a
bottleneck if file sizes are relatively small.