Introduce Inflate for database files
This function is quite simple because database files use simplified
Deflate, i.e. only one block and "static Huffman tree". It is possible
that dynamic trees will be supported in the future but the compression
ratio is acceptable at this point. Simplicity is an advantage because
trere is a lot less "surface" for bugs.