Introduce initial version of snprintf
These printf functions are not trivial and there are a lot of details
that must be taken into account. If someone familiar with C standard
libraries is going to review my code, these printf functions could be
of interest (e.g. "let's see how printf is implemented"). It is always
important to write code as accurately as possible but in this case it
is especially important. Having correct code is important, of course,
but this is like a representative of the coding conventions and there
are a lot of "competitors" because many projects have implemented the
same functions. It is possible to verify these with unit tests.

This current implementation is not complete. There are many conversion
specifiers that are not implemented at this point. I have not studied
how floating-point numbers should be handled.