Fix other minor issues
This is mostly theoretical but perhaps I should limit the buffer size
so that it would always be less than or equal to INT_MAX. By the way,
I have never used this conversion specifier n. It is very exceptional
because we are using the argument pointer which is written into. This
sounds a bit problematic and probably makes an attack surface larger.