Introduce some temporary functionality
The init executable prints the memory map. This is temporary code and
will be removed/refactored in the future. Booting the Dancy operating
system may not give much at this point but at least some information
is printed. Some developers may be interested in seeing what kind of
memory maps different computers have.

Please note that the memory map does not have "gaps" and only the base
addresses are saved. It means, by definition, that entries are sorted.