Add new release targets
There are two new release targets, usbhuge.img and usbtiny.img, but it
might be that I will disable the former. As mentioned in the previous
commit message, it is mostly for testing FAT version 16. It takes more
time to do a "make all" build because of this "huge" image file.

However, I could add some extra files, e.g. desktop background images,
that would only be available in the bigger image files but still keep
the basic system compact enough to fit in floppy disks.