Improve the linker
There have been many cases where I got an unresolved symbol error. It
would definitely be much more helpful if the linker gave the name of
the unresolved symbol. This commit introduces that functionality.

In general, I have been reading some literature. The kernel is not so
easy and some design is required up front. There is an approach called
"Big Design Up Front (BDUF)" but that definitely is not suitable for
this project. I think there should be a balance between an "improvised
hacking method" and the BDUF. Another related concept is the waterfall
model that is mostly used as an example of what not to do. However, I
should not overlook that model. Especially "software requirements" is
an important step when designing what system calls are needed.

I do have a vision of the general structure but details are subject to
change because I am learning new things all the time.