Add acpi.c file
The ACPI specification is very big and complex. Fortunately, reading
static tables is easy and there is no need to have an interpreter for
the AML bytecode. At this point, I read the "IA-PC Boot Architecture
Flags" but it is very easy to extend the code later.

Actually, as a small detail, the current boot loader has already read
the "iapc_boot_arch" internally but there is no interface for getting
it. That was on purpose because the init or kernel needs a mechanism
for getting the ACPI tables. This should be more "open" and flexible.