Detect boot loaders
The BIOS boot loader is currently implemented. It does not support an
"ah=0xE0" syscall, and that is the one of the reasons why it uses the
default number 0x00000000 in the "uint32_t boot_loader_type". I will
implement the UEFI boot loader later but I'm not in a hurry with that
because perhaps I'd like to introduce new features at the same time.
Also, I want to concentrate on the kernel and get some functionality
ready. At the same time, I get more experience and that surely helps
when designing new boot loader features.

However, probably the UEFI boot loader will be a lot simpler because
the "extensible firmware interface" already provides many features
that were relatively difficult (or at least laborious) to implement
when using assembly and the BIOS firmware.

Both the init and kernel should depend on the firmware type as little
as possible. The "uint32_t boot_loader_type" should not be used much.