Introduce a simple memory manager
Like the header says, this is a _simple_ memory manager. Simplicity is
not a guarantee that the implementation itself would be trivial. The
pre-kernel environment uses physical memory that can be accessed with
32-bit pointers. The function memory_init can be called many times for
making sure that the memory map passes validation tests even after it
has been modified by b_aligned_alloc, b_malloc, or b_free.