Clean up temporary code
The init function still has some test code but I will try keep it very
simple in the future. It could be a good starting point for checking
and studying the flow of the initialization procedures. Next I'm going
to write logging functions so that "memory maps" would not be printed
on the screen during the normal system boot.

A splash screen would be nice to have at some point. Just to make the
operating system visually appealing if nothing else.