Handle missing preprocessor defines
It is erroneous to define SIZE_MAX if the actual value is not known. I
know that writing code:

        size_t size_max = -1;

should be valid but I didn't use it because I suspected that compilers
may not always understand the intention. Please note that I'm not sure
whether there are any compilers that fail (or give some warnings) when
doing something like that. Avoiding -1 was just a precaution.