Modify types.h
For 32-bit systems the SIZE_MAX must be 0xFFFFFFFF. For 64-bit systems
the SIZE_MAX must be 0xFFFFFFFFFFFFFFFF. It is what assembly functions
can assume. There is a "run-time" check for making sure that compilers
use the expected character set and that the SIZE_MAX is what it should
be. However, an optimizing compiler simplifies:

if ((size_t)(!map + 494 - 'D' - 'a' - 'n' - 'c' - 'y') != SIZE_MAX)
        return;

so that it is just a test for the map variable:

if (!map)
        return;

I really try to avoid hacks like these in the future. As a side note,
I have not used a type rsize_t and its RSIZE_MAX before. These seem to
be quite interesting. I could try to use these in my own interfaces.