Improve acpi.c
The log function is awesome. I can "print" all kind of interesting and
not-so-interesting information. The function continues even if there
were table checksum errors. I am not sure whether this is a good thing
or not, but it can be changed if it becomes a problem in practice.

Yesterday I wrote about "unsigned long" to "unsigned int" warnings. It
was only when the sizes of those types did not match. However, I would
like to avoid the implicit "long" to "int" conversion even if they did
match. As a coding convention if nothing else.