Fix a typographic error
This is an extremely small commit but the discussion could be lengthy.
I noticed the error soon after I did the git commit. The most sensible
option would have been to postpone the modifications until I had more
contents to be added. Then these kind of errors could have been fixed
at the same time.

This commit is definitely a mistake and I should/must "let go" these
irrelevant errors. One extra letter in a comment does not make the
operating system imperfect, and because there are a massive amount of
other things that guarantee the imperfectness, it only highlights the
fact that wrong things get too much attention.