Introduce new boot loader syscalls
The BIOS boot loader is not going to implement these new system calls
and the UEFI boot loader is not implemented. This is a bit interesting
case because now I am designing interfaces for an implementation that
is not existing, and probably will not be any time soon.

It is not a coincidence that "struct b_time" is identical to EFI_TIME
as far as first six variables are concerned.