Write boot log to serial ports
The boot loader syscalls b_get_byte_com1 and b_put_byte_com1 (and com2
versions) are not very efficient. The main reason I wrote those was to
experiment and have fun. The real kernel will handle these much more
efficiently, e.g. no need to block until all data are written out. The
internal buffer size for input ("get") is not big, so if bytes are not
consumed fast enough, they are dropped.

A loop of "get one character" and "print one character on the screen"
does not work if using VBE modes, at least not on computers I have.
The combined overhead of syscalls and drawing routines means that the
buffer will overflow eventually and bytes get dropped.