Introduce rtc.c
A real time clock (RTC) uses so-called battery-backed RAM. The device
could be described as "RTC/CMOS" and it was first available in the IBM
PC/AT. Of course, there might have been a lot of embedded devices that
had it (or equivalent) but as far as "PC" computers are concerned, the
the description should be correct enough.

I used ACPI tables for checking whether the RTC was "not present", and
the TSC instruction was used for small delays between port ins/outs. I
have some kind of validation test for time values, so if the RTC did
not exist (e.g. ACPI was not available, or "not present" unsupported),
the rtc_read function should return 1. If the rtc_read returns 0, I am
quite confident that it worked.

Important note: there are two branches in the code. If the boot loader
is "BIOS", then it is assumed that a PIC device is available (whether
that is emulated or not is not relevant) and the BIOS boot loader has
used it already. If it is "non-BIOS", e.g. the UEFI boot loader, then
the code for reading RTC/CMOS shall not be used.

For time services, e.g. "GetTime", the UEFI specifications say:

   "The purpose of these interfaces is to provide operating system
    writers with an abstraction for hardware time devices, thereby
    relieving the need to access legacy hardware devices directly."

I will use this service, so the RTC/CMOS is read by the firmware.