Modify rtc.c
I tried to be careful but bugs happen. The NMI bit was interpreted in
a wrong way, so reading CMOS registers left the NMI disabled. This was
exactly the opposite of what I intended. There should not be reasons
for disabling the NMI, and it would not be something that an rtc_read
function could decide even if it were the correct strategy.