Introduce dy-conf program
The boot loader should work without user configurations. However, the
video mode must be selected every time the computer is booted. It is
possible to have a file "system/config.at" that contains some settings
that the boot loader reads.

There is a program called "dy-conf" that has a few options:

    --no-vbe-mode
        Do not use VBE video modes

    --no-vbe-menu
        Do not display user interface

    --no-hlt
        Do not use hlt instruction
        (if the computer does not support that instruction correctly)

Then there are two supported operands:

    screen_width x screen_height x bits_per_pixel (no spaces)
        Select the given video mode if it is supported

    reserved_mem_base,reserved_mem_size
        Do not use this memory range

Usage example:

    dy-conf -o system/CONFIG.AT --no-vbe-menu 800x600x16
    dy-conf -o system/CONFIG.AT 0x0000000000200000,0x0000000000100000

The former sets the video mode without showing the user interface and
the latter marks memory from 0x200000 to 0x2FFFFF as a reserved area.
In theory, it means that a computer could have a corrupted memory area
and the boot loader would not touch that. In practice, it means that I
can, as an operating system developer, use it for testing purposes.