Refactor boot loader system calls
I checked all the boot loader system calls and their return values. It
is not necessary to have any universal error value, but each function
can define it. For example, the b_get_byte_com1 function returns zero
if there are no incoming bytes. If there are, then the return value is
(byte | 0xFFFFFF00) and a simple if statement works even if the actual
byte happens to be zero.

One other advantage is that

    dy-link system/IN_IA32.AT --dump
and
    dy-link system/IN_X64.AT --dump

print actual symbol names of boot loader system calls. However, there
is a minor disadvantage: the size of the string table increased.