Modify memory.c
Now it is possible to free database memory areas. This should be done
after uncompressing them so that memory would be recycled. I made some
other changes as well. The flag B_FLAG_VALID_LEGACY might be a little
bit confusing because it includes address 0000000100000000h. There was
a logical reason for that, but probably not as relevant as I though it
would be when I defined the memory map format.