Add db.c
This component will implement all the functions that are necessary for
getting data out from the database files. The database entries in the
memory map must be parsed and then the inflate_uncompress function can
uncompress all the data. Error checking will be very important because
the boot loader does not understand the payload of the database files,
and therefore has not done any consistency checks for it.