Modify rtc.c
These modifications should be very trivial. In general, there should
be enough parentheses when using preprocessor macro parameters. Here
it is just a coding convention issue because all the macro calls can
be verified.

There was a plain typographic error, i.e. "descipher" is wrong. It is
worth noting because I explicitly checked the word before writing it!
However, the fact remains that I failed to write it correctly.