Introduce first steps of ACPICA integration
This work is in progress. I have not defined all the necessary things
in acdancy.h and the OS layer is totally unimplemented. However, I
managed to build the source code and dy-link successfully listed all
the undefined symbols, something like this:

  ...
  ...
  00000000  extern  -> _AcpiOsAcquireLock
  00000000  extern  -> _AcpiOsAcquireObject
  00000000  extern  -> _AcpiOsAllocate
  00000000  extern  -> _AcpiOsCreateCache
  00000000  extern  -> _AcpiOsCreateLock
  00000000  extern  -> _AcpiOsCreateSemaphore
  00000000  extern  -> _AcpiOsDeleteCache
  00000000  extern  -> _AcpiOsDeleteLock
  00000000  extern  -> _AcpiOsDeleteSemaphore
  00000000  extern  -> _AcpiOsEnterSleep
  00000000  extern  -> _AcpiOsExecute
  00000000  extern  -> _AcpiOsFree
  00000000  extern  -> _AcpiOsGetRootPointer
  ...
  ...

and so on. I did not enable building these acpica.o objects by default
because it makes the build process much slower. A lot of work must be
done before this is usable. However, one milestone has been reached.