Fix patch executable name
When running on a Linux environment, obviously the .exe file extension
does not work. My first test was on "MSYS2" and this script worked but
it was just a pure coincidence. A plain old "copy" error, because the
acpica.cmd was written first and the acpica.sh was a "translation".