Modify acpica.mk
Compiler-specific header files are not used. The previous commit made
them unnecessary. However, I am not absolutely sure what I am doing.
It might be that these lines must be added back in the future. All the
compilers I currently have seem to build both 32- and 64-bit versions
of the acpica.o object. It is now relatively easy to start working on
the missing symbols.

However, I have to concentrate on other aspects of the Dancy operating
system. But I like to know what features are needed for the ACPICA and
this somehow gives a good guideline of how to proceed.