Fix symbol merging
This test is for relatively unlikely scenarios. It could be that the
last "static" symbol and the first "extern" symbol have the same value
and section number. The latter symbol should not be deleted. I already
fixed this when I last modified the file but then I introduced another
bug. I noticed this because the symbol tables had two ".rdata" symbols
when the init executables were built.

This is a good example where "trivial" modifications cause new bugs.