Add more string functions
There are three new functions. Writing these is relatively easy but it
took some time to test them. The function strncpy is a bit surprising
because it can produce a "string" that is not null terminated. Perhaps
the C standard could have made the other design decision. However, I
assume that there was a good reason to do it like this.