Add helper routines
It is extremely important that these functions are bug-free. Otherwise
the whole operating system is erroneous. However, there should not be
many 64-bit div or mod operations needed on 32-bit CPUs. These helper
routines are highly inefficient so the most important optimization is
to avoid calling them in the first place. The reason I added these was
to make the "ACPI_USE_NATIVE_DIVIDE" work.

If the performance of these helper routines becomes a problem in the
future, I have to think about other optimizations.