Modify scripts
The ACPICA scripts had a bug. The git checkout did not work as I meant
because the current directory was wrong. Fortunately, the git supports
"-C <path>" that can be used for changing the current directory. "Run
as if git was started in <path>", to quote its documentation.

In general, programs should have a standardized way of doing this. I
think scripts should not change the current directory unless there was
a really good reason to do so.