Rename boot log functions
I made a big mistake when using the name "log". The C standard library
already has "double log(double x)". Although the header "math.h" isn't
used in the kernel, I really want to avoid using any reserved function
names.

The compilers might handle this "log" in a special way because it is a
very good candidate for a builtin function. Fortunately this was fixed
early enough and the new prefix is "b_log" from now on.