Modify types.h
If the SIZE_MAX is defined by compiler, use that value. All assembly
functions assume that sizeof(size_t) is 4 for 32-bit and 8 for 64-bit.

For example:

            ; void *memset(void *s, int c, size_t n)
    _memset:
            push edi                        ; save register edi
            mov edi, [esp+8]                ; edi = s
            mov eax, [esp+12]               ; eax = c (low byte)
            mov ecx, [esp+16]               ; ecx = n
            ...
            ...

The above example shows that "n" must be a 32-bit value. And if that
is not true, the whole ABI breaks and my assembly functions behave in
a very unpredictable way.