Build database files
There are only two files that are packed into a set of "DB" files. The
ACPICA is not usable at this point but it can be used as a payload. I
need to write code for handling the database files. The implementation
will be in file "db.c" which is empty now.