Require support for unsigned long long
The 32-bit Dancy must be built with a compiler that supports "unsigned
long long" (or in general, "long long"). The ACPICA code base already
requires this, so I think it would make sense to harmonize this aspect
of the tool chain. This decision simplifies a lot of things because I
can use a uint64_t type. However, on 32-bit platforms, it is not very
cheap to have complicated operations, e.g. 64-bit divisions.

It might be that I have to add a few more "helper routines" but I try
to keep the usage of those at absolute minimum. Both for performance
and elegancy reasons but correctness is also one concern.