Modify acpica scripts
The reason for this change is to make sure that "modified" timestamps
are always updated whenever the files are copied. At the same time, I
thought that I should keep the "external" directory more standard, so
"acpica" was moved to "tmp". Having it in "tmp" hightlights the fact
that those files should not be modified.

All test modifications or similar must be done on files that are in
"include/acpica" or "kernel/acpica". Then the build system should work
as it should, although modifications to include files always requires
a full build. This is not optimal but simplifies dependencies.