Add helper routines
Like with the previous commit that added new "helper routines", I have
not optimized these at all. The most important optimization is still
to avoid calling these helper routines in the first place. It might be
that I will not need much more of these because the ACPICA code base
was the "external" reason why I needed most of them.