Add proper intmax_t support
A modern snprintf must support "intmax_t" and "uintmax_t" types. The
C11 draft standard says,

   "Specifies that a following d, i, o, u, x, or X conversion
    specifier applies to an intmax_t or uintmax_t argument; or
    that a following n conversion specifier applies to a pointer
    to an intmax_t argument."

for the length modifier "j". I think that now my snprintf has enough
features implemented for the kernel. I try to check how difficult it
is to add support for floating-point numbers, although this shouldn't
be a must-have feature. However, if this implementation was otherwise
almost complete, this part could be added for full completeness.