Build nasm as part of external tools
The Netwide Assembler (NASM) has always worked and I have been using
it on DOS, Windows, and Linux for many years. In the future, I may
write my own assembler but the syntax and program interface will be
the same as in NASM. My version would not have that many features but
a strict subset of features should be enough. My assembly routines in
the Dancy project are quite simple as far as assembler requirements
are concerned. Preprocessor usage is very minimum.

Hopefully my build script is correct. I checked "Beyond Linux From
Scratch" project and they are building the NASM with a simple set
of commands: "./configure && make && make install". Of course, the
"--prefix" is given to the configure script. My installation does
not require root privileges because the target is "external/bin".