Modify limits.h
The rules for data type sizes are the following:

    "unsigned char"         1 byte
    "unsigned short"        2 bytes
    "unsigned int"          4 bytes
    "unsigned long"         4 bytes      (Dancy 32)
                            4 or 8 bytes (Dancy 64)
    "unsigned long long"    8 bytes

If these are something else, the compiler is not compatible. However,
the coding convention is to NOT assume these exact sizes and this is a
very important thing to note specifically. The "snprintf" has assumed
the exact sizes but there is a preprocessor "if" that prevents it from
compiling if "DANCY_32" or "DANCY_64" both are undefined.

Types "uint8_t", "uint16_t", "uint32_t", or "uint64_t" must be used if
exact sizes are needed, e.g. in structures.

However, it is possible to assume that "int" is 4 bytes or more, e.g.
"int i = 65536" is valid, when programming the Dancy kernel.