Refactor init.h
There are a lot of .c files that need only a very few declarations in
the header. That is why I combined the whole "init" in one header file
but the file was not very readable. This commit should make it better.
First I tried to use only the file names as comments, e.g.

    /*
     * init.c
     */
    void init(void *map);

but the comments were not "highlighted" enough. I added a more verbose
comment "Declarations of init.c" to make it more visible.