Add UEFI support
It has been clear that BIOS is a legacy firmware and UEFI support must
be implemented. Today I tried a "hello world" EFI/BOOT/BOOTX64.EFI and
it was booted from "dancy.iso" that was built like this:

    ...
    dy-mcopy -i temp.img ./BOOTX64.EFI ::EFI/BOOT/BOOTX64.EFI
    dy-iso -o dancy.iso temp.img --uefi

The UEFI specification says:

   "To boot from a CD-ROM or DVD-ROM in the boot services environment,
    an EFI System partition is stored in a "no emulation" mode as
    defined by the "El Torito" specification. A Platform ID of 0xEF
    indicates an EFI System Partition."

Because I have had DANCY.IMG all the time, the EFI System partition
is just the same image as for the BIOS boot loader.