Introduce database functions
The function "db_read" is the most important. It returns a pointer to
an allocated memory area that contains the uncompressed data. The size
of the data is returned as a second value. Currently there are no more
than two files "o32/acpica.o" and "o64/acpica.o" that can be read. It
is very easy add more files later.

The caller is responsible for freeing the allocated buffers. There are
no strict alignment requirements (i.e. only what malloc provides) so
the data will probably be copied to another buffer after it has been
processed. This is specifically the case with object files, e.g. those
acpica.o sections will be copied into page-aligned memory areas and
then all the relocations are applied, among other things.

I will write another set of functions (file "ld.c") that handles the
things mentioned above. Dynamic linking is an interesting topic that I
have to study more, e.g. whether I want it or not.