Add object validation function
The object file format is very strict. It has four sections: ".text",
".rdata", ".data", and ".bss". It means that only files produced by
the dy-link program are valid. The main advantage of using a standard
format is that there are a lot of third party tools that understand
it, e.g. disassemblers.

I took some code from my dy-link program. I'm not extremely happy with
this code quality. Fortunately, the "ld_validate" function does not do
any modifications. It should be treated like a black box that answers
to a question "is this object file valid". In general, this test isn't
extremely critical because the data should be valid.

One thing worth mentioning is that so called common symbols need not
to be handled because the linker has already allocated space in the
.bss section. Also, the symbol table does not have "extra records" so
it can be "accessed by index".