Fix the CPU index test
There was an obvious logic error. This error is a good example of how
coding conventions are important. My convention is to put the variable
to be tested first, e.g. "if (var > 3)" and NOT "if (3 < var)". Here I
accidentally put "core.cpu_max" first and then used "greater than" as
in the example above. It is more common to have tests for checking if
variable is in certain limits and "if (true)" is an error condition.

You may wonder why I write these kind of commit messages for a trivial
modification. This helps avoiding similar errors because it takes some
time to think the problem and _it_could_have_been_a_major_bug. In this
case it was not a major bug at all because it would have been detected
(correct arguments would have failed) and, in general, it's not likely
that a caller provides the function with invalid arguments.