Update copyright year
I have not developed the operating system for months. Now it is a good
time to concentrate on this interesting task and, first of all, I have
to implement a UEFI boot loader because the BIOS firmware is going to
be outdated. However, it is very good that I have the BIOS boot loader
because old computers will "always" be here and there (and emulators).