Develop the UEFI boot loader
Programming in C is a lot easier than programming in assembly. I have
written only those UEFI definitions that are needed at the moment. By
the way, having a printf-like function is like a milestone because it
makes the (in)famous printf-debugging techniques available.

I will try to release a "dancy.iso" image for testing purposes.