Develop the UEFI boot loader
This commit touches many source code files. I wrote memory functions
but those are not fully implemented yet. Especially, a function called
"memory_export_map" will be time-consuming because it must be able to
update the native map (the init executable might have modified it). It
is important to note that the UEFI firmware owns the memory map before
"ExitBootServices()" is called. The b_exit function (init syscall) is
the one that will eventually call it.

When the init allocates "free memory", it will use an EfiLoaderData
memory area. Part of the solution is the "B_FLAG_NO_INIT_ALLOC" flag.
With that it is possible to detect "real free memory" and "simulated
free memory" (type EfiLoaderData).

The reason for complexity is that the UEFI firmware might modify the
memory map when using boot service calls. It must be the _latest_ map
that will be given to the operating system, so the final boot service
calls are "GetMemoryMap()" and "ExitBootServices()", in that order.