Use const type qualifiers
In this case this is not very important, but because I already added
const qualifiers for some map pointers. I thought it would make sense
to harmonize coding conventions. By the way, the (const void *) casts
are just to get rid of compiler warnings.

    const unsigned char *map;
    const EFI_MEMORY_DESCRIPTOR *entry;

    entry = (const void *)map;

chars do not have the same alignment requirement, but this tells the
compiler that everything is OK.