Use AllocateAddress instead of AllocateMaxAddress
I totally ignored a huge problem. It would have been a nightmare to
handle scenarios where the "Attribute" value is not the same for the
allocated contiguous memory block. For example:

Memory Map
    000000001FF80000 000000001FF9FFFF  Executable (Init)
    000000001FFA0000 000000001FFA0FFF  Database (0)
    000000001FFA1000 000000001FFA1FFF  Database (0)     (!)
    000000001FFA2000 000000001FFAFFFF  Database (0)     (!)
    000000001FFB0000 000000001FFBFFFF  Database (1)
    000000001FFC0000 000000001FFCFFFF  Database (2)

Here the database 0 has been splitted into three parts. I solved the
problem so that the allocated memory must have the same "Attribute".
Theoretically this is not 100% safe because the UEFI firmware could
change the capabilities when allocating memory blocks.