Introduce file functions
The fundamental function of a "loader" is to load the operating system
files into memory. Although the file.c implementation looks a bit more
complicated than what I assumed, it is a lot simpler than the old BIOS
boot loader. No need to understand the FAT file system, use "int 0x13"
disk services, 16/32 bit transitions, assembly language...

However, the error messages are not extremely detailed and I did not
retry the file reading functions if they failed (EFI_SUCCESS != 0). It
is assumed that the UEFI firmware does its job when it comes to that.
The CRC-32C checksums should be quite reliable safety nets so all the
files having the checksum (config.at and in_x64.at) are likely valid.
The databases do not have that because the payload itself has its own
checksum mechanisms.

For all the files, the native file headers are checked.