Implement b_get_parameter function
There are two guids for the "B_ACPI_POINTER", the EFI_ACPI_TABLE_GUID
and ACPI_TABLE_GUID. The former is searched first and used if its RSDP
address is 32-bit-compatible. I have to modify the interface of the
boot syscall if I need to return a 64-bit value (unsigned long is not
necessarily 64-bit). It is doable, but I think that it would not make
sense to have a 64-bit RSDP address.

I will modify the init executable so that computers with UEFI firmware
require valid ACPI tables. It will handle the unlikely case that the
RSDP pointer is not returned because of the 64-bit address value.