Set key toggle bits
The NUMLOCK, CAPSLOCK, and SCROLLLOCK are added. The BIOS boot loader
has these same bits and now the functionality is ready. However, it is
not possible to write a good keyboard driver based these boot services
and it is expected that only some keys will be actually used, e.g. F1,
F2, F3, Escape, arrow keys, and so on. This mostly solves the keyboard
layout problem that would be too complicated for the init executable.