Refactor and add comments
The "uefi_main" source file has found its shape. Any changes to this
file are probably very minor. The UEFI boot loader is still missing
the video mode selection menu but otherwise the main functionality of
the loader is ready. It was a lot easier than the BIOS boot loader,
but the comparison is not very fair. The UEFI firmware is like an OS
that provided a huge framework, e.g. a file system driver.