Accept external symbols
The native output format (-f at) accepts external symbols. This means
that it is possible to have dynamically linked executables. I have to
think whether there should be a method for limiting what symbols are
really "exported" because not all external symbols are that important
or safe to be called from other modules.