Reorganize init source code
The init executable has two parts. The second part is read from the
the system databases and this solves the problem with the file sizes
of IN_IA32.AT and IN_X64.AT.

The native object "init.at" is compressed so there are practically no
size limits (assuming that there is enough free memory). It will be
possible to have some graphics and proper fonts, for example.