Add db_read_error function
This error handling strategy is not very good. I have to introduce a
set of error codes and their defines in some kind of header file. The
db_read function did not print anything because I thought that there
might be situations in which the init just checks whether a file can
be read or not. For example, try to set the background image, etc.