Develop dy-ttf
The glyf table contains "simple glyphs" and "compound glyphs". I wrote
code that can extract coordinate values from the simple glyphs but I
have not verified the results. I need to visualize the lines somehow,
and at first it is not even necessary to evaluate the Bezier curves.
Simple straight lines would "visualize" the overall shape of the font.