Develop dy-ttf
I forgot to check and fix the compiler warnings, and this should be a
basic thing before creating a new commit. However, the program is not
ready, but it can do some interesting things (for learning purposes).

For example, on Windows computers:

dy-ttf.exe -o A.bmp --render --verbose c:\Windows\Fonts\arial.ttf 0x41

It creates a bitmap, A.bmp, and prints this:

  <TTGlyph name="A" xMin="-3" yMin="0" xMax="1369" yMax="1466">
    <contour>
      <pt x="-3" y="0" on="1"/>
      <pt x="560" y="1466" on="1"/>
      <pt x="769" y="1466" on="1"/>
      <pt x="1369" y="0" on="1"/>
      <pt x="1148" y="0" on="1"/>
      <pt x="977" y="444" on="1"/>
      <pt x="364" y="444" on="1"/>
      <pt x="203" y="0" on="1"/>
    </contour>
    <contour>
      <pt x="420" y="602" on="1"/>
      <pt x="917" y="602" on="1"/>
      <pt x="764" y="1008" on="1"/>
      <pt x="694" y="1193" on="0"/>
      <pt x="660" y="1312" on="1"/>
      <pt x="632" y="1171" on="0"/>
      <pt x="581" y="1032" on="1"/>
    </contour>
  </TTGlyph>

There are no Bezier curves. Only the "Bresenham's line algorithm" was
used, very inefficiently. Green points are "on curve" and red points
are the control points.