Develop dy-ttf
It is very obvious that "downscaling" is not simple. My simple scaling
for the "glyph renderer" is just for testing purposes. If the option
"-s divisor" option is given, only the "glyph graphics" is rendered on
the bitmap. I think that I have learnt the basics of the TrueType file
format and can extract relevant data from it.