Implement ttf_build functions
This commit introduces many "ttf_build" functions. Now this program
can create simplified TrueType files but I assume that I have to fine
adjust some features when fonts are actually used. Reading the "glyf"
table should be much simpler because an array of flags contains only
values 0 or 1 ("off-curve" and "on-curve"). No need to keep track of
"repeat" states, "short vectors", and so on. This increases the file
size but I think that this should be acceptable.

However, functions that consume "Dancy TTF" files must check the data
validity. An error message, e.g. "incompatible format", can be used.