Modify ttf.c
It was not just a coding convention to put "0u" instead of "0". Some
compiler versions display a warning message because the return value
type is size_t (that is unsigned). Plain "return 0" would not do that
but the comma operator makes it more "complicated" to the compiler.

I like the "return free(something), 0" idiom because it is quite easy
to quickly check all the returns and make sure that "free()" or some
other procedure is always run before exiting the function.