Modify ttf.c
There was a bug in my buffer size check. Because the TTF file format
uses 16-bit integers quite extensively, most of the overflow checks
are quite easy with 32-bit types.

In general, it is very easy to have subtle overflow bugs. One thing
I miss from assembly is the "carry flag", e.g.

        add eax, [value1]
        jc handle_error
        add eax, [value2]
        jc handle_error
        ...

This kind of programming is very "linear", of course, but it is quite
simple to handle overflows. In C, these kind of checks are "advanded"
topics in programming books although they should be (almost) the very
first lessons after an initial "hello world" example.